requirements.txt
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
future==0.18.2
imutils==0.5.4
numpy==1.23.4
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.5.1
pkg_resources==0.0.0
python-dateutil==2.8.2
python-vlc==3.0.16120
pytz==2022.5
six==1.16.0
